Decompile matching sound-effect playback and control

+++ src/sound_bgm.c
@@ -12,7 +12,7 @@ void sub_02005EEC(void);
 void NNS_SndPlayerStopSeqBySeqNo(int seqNo, int fadeFrames);
 int GF_GetPlayerNoBySeq(u16 seqNo);
 int sub_02005328(int state);
-void sub_02005464(u16 seqNo, enum SoundHandleNo handleNo);
+void sub_02005464(int seqNo, enum SoundHandleNo handleNo);
 void sub_020059E0(u16 seqNo);
 BOOL sub_02005D10(u16 seqNo);
 void sub_02005DA0(u16 seqNo, enum SoundHandleNo handleNo);
